fix: period recruitment test구분 및 에러코드 수정
